Implement strategy pattern for cleaning. #3
