+++ src/main/java/de/hpi/urlcleaner/service/IUrlCleanerService.java
@@ -2,7 +2,6 @@ package de.hpi.urlcleaner.service;
 
 public interface IUrlCleanerService {
 
-    String cleanTrackers(String dirtyUrl);
+    String clean(String dirtyUrl, long shopID);
 
-    String cleanRedirects(long shopID, String url);
 }
