+++ src/main/java/de/hpi/urlcleaner/service/core/ICleanStrategy.java
@@ -1,0 +1,6 @@
+package de.hpi.urlcleaner.service.core;
+
+public interface ICleanStrategy {
+
+    String clean(String dirtyUrl);
+}
