+++ src/test/java/de/hpi/urlcleaner/service/core/TrackerCleanStrategyTest.java
@@ -1,4 +1,4 @@
-package de.hpi.urlcleaner.service;
+package de.hpi.urlcleaner.service.core;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AccessLevel;
@@ -13,13 +13,13 @@ import static org.junit.Assert.assertEquals;
 
 @Getter(AccessLevel.PRIVATE)
 @Setter(AccessLevel.PRIVATE)
-public class UrlCleanerServiceTest {
+public class TrackerCleanStrategyTest {
 
-    private IUrlCleanerService service;
+    private ICleanStrategy strategy;
 
     @Before
     public void setup() throws IOException {
-        setService(new UrlCleanerService(new ObjectMapper()));
+        setStrategy(new TrackerCleanStrategy(new ObjectMapper()));
     }
 
     @Test
@@ -28,7 +28,7 @@ public class UrlCleanerServiceTest {
                 ".de/tintenpatrone-schwarz-kompatibel-11ml-mit-chip-1426744a?sPartner=idealo";
         String expectedUrl = "https://www.silvertrade" +
                 ".de/tintenpatrone-schwarz-kompatibel-11ml-mit-chip-1426744a";
-        assertEquals(expectedUrl, getService().cleanTrackers(dirtyUrl));
+        assertEquals(expectedUrl, getStrategy().clean(dirtyUrl));
     }
 
     @Test
@@ -37,7 +37,7 @@ public class UrlCleanerServiceTest {
                 ".html?utm_source=web&utm_medium=cpc&utm_campaign=idealo";
         String expectedUrl = "https://www.aze-tuning.de/sitzbezug-sitzbezuege-mitsubishi-carisma" +
                 ".html";
-        assertEquals(expectedUrl, getService().cleanTrackers(dirtyUrl));
+        assertEquals(expectedUrl, getStrategy().clean(dirtyUrl));
     }
 
     @Test
@@ -45,7 +45,7 @@ public class UrlCleanerServiceTest {
         String dirtyUrl = "https://www.mega-bikes" +
                 ".de/krg-shimano-ultegra-3953-zaehne-1725mm-fc-6800-hollowtech-ii-m-achse-11-fach:15144.html?ref=6&rabatt=true";
         String expectedUrl = "https://www.mega-bikes.de/krg-shimano-ultegra-3953-zaehne-1725mm-fc-6800-hollowtech-ii-m-achse-11-fach:15144.html?rabatt=true";
-        assertEquals(expectedUrl, getService().cleanTrackers(dirtyUrl));
+        assertEquals(expectedUrl, getStrategy().clean(dirtyUrl));
     }
 
     @Test
@@ -54,20 +54,20 @@ public class UrlCleanerServiceTest {
                 ".de/krg-shimano-ultegra-3953-zaehne-1725mm-fc-6800-hollowtech-ii-m-achse-11-fach:15144.html?utm_medium=cpc&ref=6" +
                 "&rabatt=true&utm_campaign=idealo";
         String expectedUrl = "https://www.mega-bikes.de/krg-shimano-ultegra-3953-zaehne-1725mm-fc-6800-hollowtech-ii-m-achse-11-fach:15144.html?rabatt=true";
-        assertEquals(expectedUrl, getService().cleanTrackers(dirtyUrl));
+        assertEquals(expectedUrl, getStrategy().clean(dirtyUrl));
     }
 
     @Test
     public void testMixedDelimiters() {
         String dirtyUrl = "http://www.amazon.de/dp/B00UI7XAJ8/?smid=A12ZUNPYXJYAM9";
         String expectedUrl = "http://www.amazon.de/dp/B00UI7XAJ8/";
-        assertEquals(expectedUrl, getService().cleanTrackers(dirtyUrl));
+        assertEquals(expectedUrl, getStrategy().clean(dirtyUrl));
     }
 
     @Test
     public void testUpperCaseUntouched() {
         String dirtyUrl = "http://www.amazon.de/dp/B00UI7XAJ8";
         String expectedUrl = "http://www.amazon.de/dp/B00UI7XAJ8";
-        assertEquals(expectedUrl, getService().cleanTrackers(dirtyUrl));
+        assertEquals(expectedUrl, getStrategy().clean(dirtyUrl));
     }
 }
